Fit Multivariate Meta-Analysis

This function estimates the mean and covariance matrix of a vector of coefficients using the estimated coefficients and sampling variance-covariance matrix from each individual.

Usage

Meta(
  y,
  v,
  mu_start = NULL,
  mu_lbound = NULL,
  mu_ubound = NULL,
  sigma_l_start = NULL,
  sigma_l_lbound = NULL,
  sigma_l_ubound = NULL,
  diag = FALSE,
  try = 1000,
  ncores = NULL,
  ...
)

Arguments

y

A list. Each element of the list is a numeric vector of estimated coefficients.

v

A list. Each element of the list is a sampling variance-covariance matrix of y.

mu_start

Numeric vector. Optional vector of starting values for mu.

mu_lbound

Numeric vector. Optional vector of lower bound values for mu.

mu_ubound

Numeric vector. Optional vector of upper bound values for mu.

sigma_l_start

Numeric matrix. Optional matrix of starting values for t(chol(sigma)).

sigma_l_lbound

Numeric matrix. Optional matrix of lower bound values for t(chol(sigma)).

sigma_l_ubound

Numeric matrix. Optional matrix of upper bound values for t(chol(sigma)).

diag

Logical. If diag = TRUE, sigma is a diagonal matrix. If diag = FALSE, sigma is a symmetric matrix.

try

Positive integer. Number of extra optimization tries.

ncores

Positive integer. Number of cores to use.

...

Additional optional arguments to pass to mxTryHardctsem.

Details

For \(i = \left\{1, \cdots, n \right\}\), the objective function used to estimate the mean \(\boldsymbol{\mu}\) and covariance matrix \(\boldsymbol{\Sigma}\) of the random coefficients \(\mathbf{y}_{i}\) is given by $$ \ell \left( \boldsymbol{\mu} , \boldsymbol{\Sigma} \mid \mathbf{y}_{i}, \mathbb{V} \left( \mathbf{y}_{i} \right) \right) = - \frac{1}{2} \left[ q \log \left( 2 \pi \right) + \log \left( \left| \mathbb{V} \left( \mathbf{y}_{i} \right) - \boldsymbol{\Sigma} \right| \right) + \left( \mathbf{y}_{i} - \boldsymbol{\mu} \right)^{\prime} \left( \mathbb{V} \left( \mathbf{y}_{i} \right) - \boldsymbol{\Sigma} \right)^{-1} \left( \mathbf{y}_{i} - \boldsymbol{\mu} \right) \right] $$ where \(q\) is the number of unique elements in \(\boldsymbol{\mu}\) and \(\boldsymbol{\Sigma}\), and \(\mathbb{V} \left( \mathbf{y}_{i} \right)\) is the sampling variance-covariance matrix of \(\mathbf{y}_{i}\).

References

Neale, M. C., Hunter, M. D., Pritikin, J. N., Zahery, M., Brick, T. R., Kirkpatrick, R. M., Estabrook, R., Bates, T. C., Maes, H. H., & Boker, S. M. (2015). OpenMx 2.0: Extended structural equation and statistical modeling. Psychometrika, 81(2), 535–549. doi:10.1007/s11336-014-9435-8

See also

Author

Ivan Jacob Agaloos Pesigan