Fit Multivariate Meta-Analysis

This function estimates the mean and covariance matrix of a vector of coefficients using the estimated coefficients and sampling variance-covariance matrix from each individual.

Usage

MetaVARMx(
  object,
  mu_start = NULL,
  mu_lbound = NULL,
  mu_ubound = NULL,
  sigma_l_start = NULL,
  sigma_l_lbound = NULL,
  sigma_l_ubound = NULL,
  diag = FALSE,
  intercept = FALSE,
  noise = FALSE,
  error = FALSE,
  try = 1000,
  ncores = NULL,
  ...
)

Arguments

object

Output of the fitDTVARMx::FitDTVARIDMx() or fitCTVARMx::FitCTVARIDMx() functions.

mu_start

Numeric vector. Optional vector of starting values for mu.

mu_lbound

Numeric vector. Optional vector of lower bound values for mu.

mu_ubound

Numeric vector. Optional vector of upper bound values for mu.

sigma_l_start

Numeric matrix. Optional matrix of starting values for t(chol(sigma)).

sigma_l_lbound

Numeric matrix. Optional matrix of lower bound values for t(chol(sigma)).

sigma_l_ubound

Numeric matrix. Optional matrix of upper bound values for t(chol(sigma)).

diag

Logical. If diag = TRUE, sigma is a diagonal matrix. If diag = FALSE, sigma is a symmetric matrix.

intercept

Logical. If intercept = TRUE, include estimates of the process intercept vector, if available. If intercept = FALSE, exclude estimates of the process intercept vector.

noise

Logical. If noise = TRUE, include estimates of the process noise matrix, if available. If noise = FALSE, exclude estimates of the process noise matrix.

error

Logical. If error = TRUE, include estimates of the measurement error matrix, if available. If error = FALSE, exclude estimates of the measurement error matrix.

try

Positive integer. Number of extra optimization tries.

ncores

Positive integer. Number of cores to use.

...

Additional optional arguments to pass to mxTryHard.

Details

For \(i = \left\{1, \cdots, n \right\}\), the objective function used to estimate the mean \(\boldsymbol{\mu}\) and covariance matrix \(\boldsymbol{\Sigma}\) of the random coefficients \(\mathbf{y}_{i}\) is given by $$ \ell \left( \boldsymbol{\mu} , \boldsymbol{\Sigma} \mid \mathbf{y}_{i}, \mathbb{V} \left( \mathbf{y}_{i} \right) \right) = - \frac{1}{2} \left[ q \log \left( 2 \pi \right) + \log \left( \left| \mathbb{V} \left( \mathbf{y}_{i} \right) - \boldsymbol{\Sigma} \right| \right) + \left( \mathbf{y}_{i} - \boldsymbol{\mu} \right)^{\prime} \left( \mathbb{V} \left( \mathbf{y}_{i} \right) - \boldsymbol{\Sigma} \right)^{-1} \left( \mathbf{y}_{i} - \boldsymbol{\mu} \right) \right] $$ where \(q\) is the number of unique elements in \(\boldsymbol{\mu}\) and \(\boldsymbol{\Sigma}\), and \(\mathbb{V} \left( \mathbf{y}_{i} \right)\) is the sampling variance-covariance matrix of \(\mathbf{y}_{i}\).

References

Neale, M. C., Hunter, M. D., Pritikin, J. N., Zahery, M., Brick, T. R., Kirkpatrick, R. M., Estabrook, R., Bates, T. C., Maes, H. H., & Boker, S. M. (2015). OpenMx 2.0: Extended structural equation and statistical modeling. Psychometrika, 81(2), 535–549. doi:10.1007/s11336-014-9435-8

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

if (FALSE) { # \dontrun{
# Generate data using the simStateSpace package------------------------------
beta_mu <- matrix(
  data = c(
    0.7, 0.5, -0.1,
    0.0, 0.6, 0.4,
    0, 0, 0.5
  ),
  nrow = 3
)
beta_sigma <- diag(3 * 3)
beta <- simStateSpace::SimBetaN(
  n = 5,
  beta = beta_mu,
  vcov_beta_vec_l = t(chol(beta_sigma))
)
sim <- simStateSpace::SimSSMVARIVary(
  n = 5,
  time = 100,
  mu0 = list(rep(x = 0, times = 3)),
  sigma0_l = list(t(chol(diag(3)))),
  alpha = list(rep(x = 0, times = 3)),
  beta = beta,
  psi_l = list(t(chol(diag(3))))
)
data <- as.data.frame(sim)

# Fit the model--------------------------------------------------------------
library(fitDTVARMx)
fit <- FitDTVARIDMx(
  data = data,
  observed = c("y1", "y2", "y3"),
  id = "id"
)
# Multivariate meta-analysis-------------------------------------------------
library(metaVAR)
meta <- MetaVARMx(fit)
print(meta)
summary(meta)
coef(meta)
vcov(meta)
} # }