Fit Multivariate Meta-Analysis

This function estimates fixed-, random-, or mixed-effects meta-analysis parameters using the estimated coefficients and sampling variance-covariance matrix from each individual fitted using the fitDTVARMx::FitDTVARIDMx() function.

Usage

MetaVARMx(
  object,
  x = NULL,
  alpha_values = NULL,
  alpha_free = NULL,
  alpha_lbound = NULL,
  alpha_ubound = NULL,
  beta_values = NULL,
  beta_free = NULL,
  beta_lbound = NULL,
  beta_ubound = NULL,
  tau_values = NULL,
  tau_free = NULL,
  tau_lbound = NULL,
  tau_ubound = NULL,
  random = TRUE,
  diag = FALSE,
  intercept = FALSE,
  noise = FALSE,
  error = FALSE,
  try = 1000,
  ncores = NULL,
  ...
)

Arguments

object

Output of the fitDTVARMx::FitDTVARIDMx() function.

x

An optional list. Each element of the list is a numeric vector of covariates for the mixed-effects model.

alpha_values

Numeric vector. Optional vector of starting values for alpha.

alpha_free

Logical vector. Optional vector of free (TRUE) parameters for alpha.

alpha_lbound

Numeric vector. Optional vector of lower bound values for alpha.

alpha_ubound

Numeric vector. Optional vector of upper bound values for alpha.

beta_values

Numeric matrix. Optional matrix of starting values for beta.

beta_free

Logical matrix. Optional matrix of free (TRUE) parameters for beta.

beta_lbound

Numeric matrix. Optional matrix of lower bound values for beta.

beta_ubound

Numeric matrix. Optional matrix of upper bound values for beta.

tau_values

Numeric matrix. Optional matrix of starting values for t(chol(tau_sqr)).

tau_free

Numeric matrix. Optional matrix of free (TRUE) parameters for t(chol(tau_sqr)).

tau_lbound

Numeric matrix. Optional matrix of lower bound values for t(chol(tau_sqr)).

tau_ubound

Numeric matrix. Optional matrix of upper bound values for t(chol(tau_sqr)).

random

Logical. If random = TRUE, estimates random effects. If random = FALSE, tau_sqr is a null matrix.

diag

Logical. If diag = TRUE, tau_sqr is a diagonal matrix. If diag = FALSE, tau_sqr is a symmetric matrix.

intercept

Logical. If intercept = TRUE, include estimates of the process intercept vector, if available. If intercept = FALSE, exclude estimates of the process intercept vector.

noise

Logical. If noise = TRUE, include estimates of the process noise matrix, if available. If noise = FALSE, exclude estimates of the process noise matrix.

error

Logical. If error = TRUE, include estimates of the measurement error matrix, if available. If error = FALSE, exclude estimates of the measurement error matrix.

try

Positive integer. Number of extra optimization tries.

ncores

Positive integer. Number of cores to use.

...

Additional optional arguments to pass to mxTryHardctsem.

References

Cheung, M. W.-L. (2015). Meta-analysis: A structural equation modeling approach. Wiley. doi:10.1002/9781118957813

Neale, M. C., Hunter, M. D., Pritikin, J. N., Zahery, M., Brick, T. R., Kirkpatrick, R. M., Estabrook, R., Bates, T. C., Maes, H. H., & Boker, S. M. (2015). OpenMx 2.0: Extended structural equation and statistical modeling. Psychometrika, 81(2), 535–549. doi:10.1007/s11336-014-9435-8

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

if (FALSE) { # \dontrun{
# Generate data using the simStateSpace package------------------------------
beta_mu <- matrix(
  data = c(
    0.7, 0.5, -0.1,
    0.0, 0.6, 0.4,
    0, 0, 0.5
  ),
  nrow = 3
)
beta_sigma <- diag(3 * 3)
beta <- simStateSpace::SimBetaN(
  n = 5,
  beta = beta_mu,
  vcov_beta_vec_l = t(chol(beta_sigma))
)
sim <- simStateSpace::SimSSMVARIVary(
  n = 5,
  time = 100,
  mu0 = list(rep(x = 0, times = 3)),
  sigma0_l = list(t(chol(diag(3)))),
  alpha = list(rep(x = 0, times = 3)),
  beta = beta,
  psi_l = list(t(chol(diag(3))))
)
data <- as.data.frame(sim)

# Fit the model--------------------------------------------------------------
library(fitDTVARMx)
fit <- FitDTVARIDMx(
  data = data,
  observed = c("y1", "y2", "y3"),
  id = "id"
)

# Multivariate meta-analysis-------------------------------------------------
library(metaVAR)
meta <- MetaVARMx(fit)
print(meta)
summary(meta)
coef(meta)
vcov(meta)
} # }